+++ tests/conftest.py
@@ -1,19 +1,46 @@
 import os
 
 import pytest
+from sqlalchemy import create_engine
 
 
-@pytest.fixture
-def mysql_url():
+@pytest.fixture(scope='module')
+def mysql_engine():
     try:
-        return os.environ['MYSQL_URL']
+        url = os.environ['MYSQL_URL']
     except KeyError:
         pytest.skip('MYSQL_URL environment variable is not set')
+    else:
+        engine = create_engine(url)
+        yield engine
+        engine.dispose()
 
 
-@pytest.fixture
-def postgresql_url():
+@pytest.fixture(scope='module')
+def postgresql_engine():
     try:
-        return os.environ['POSTGRESQL_URL']
+        url = os.environ['POSTGRESQL_URL']
     except KeyError:
         pytest.skip('POSTGRESQL_URL environment variable is not set')
+    else:
+        engine = create_engine(url)
+        yield engine
+        engine.dispose()
+
+
+@pytest.fixture(scope='module')
+def sqlite_memory_engine():
+    engine = create_engine('sqlite:///:memory:', connect_args=dict(check_same_thread=False))
+    yield engine
+    engine.dispose()
+
+
+@pytest.fixture(scope='module')
+def sqlite_file_engine(tmpdir_factory):
+    tmpdir = tmpdir_factory.mktemp('asphalt-sqlalchemy')
+    db_path = tmpdir.join('test.db')
+    engine = create_engine('sqlite:///%s' % db_path, connect_args=dict(check_same_thread=False))
+    yield engine
+    engine.dispose()
+    if db_path.exists():
+        db_path.remove()